color change in extra file, kochsnowflake beginning
